fix(FON-24): can provide outcomes to multiple nomination files
