+++ apps/client/src/components/secretariat-general/transparence/tableau-affectation-dossier-de-nomination/nomination-file-outcome/NominationFileOutcomeBadge.tsx
